+++ modules/AWSPowerShell/AWSAliases.ps1
@@ -3344,6 +3344,13 @@ Set-Alias -Name EMS-GetContainerPolicy -Value Get-EMSContainerPolicy
 Set-Alias -Name EMS-ListContainers -Value Get-EMSContainerList
 Set-Alias -Name Put-EMSContainerPolicy -Value Write-EMSContainerPolicy
 Set-Alias -Name EMS-PutContainerPolicy -Value Write-EMSContainerPolicy
+Set-Alias -Name Delete-EMSDObject -Value Remove-EMSDObject
+Set-Alias -Name EMSD-DeleteObject -Value Remove-EMSDObject
+Set-Alias -Name Describe-EMSDObjectMetadata -Value Get-EMSDObjectMetadata
+Set-Alias -Name EMSD-DescribeObject -Value Get-EMSDObjectMetadata
+Set-Alias -Name EMSD-GetObject -Value Get-EMSDObject
+Set-Alias -Name EMSD-ListItems -Value Get-EMSDItemList
+Set-Alias -Name EMSD-PutObject -Value Write-EMSDObject
 Set-Alias -Name MH-AssociateCreatedArtifact -Value Add-MHCreatedArtifact
 Set-Alias -Name MH-AssociateDiscoveredResource -Value Add-MHDiscoveredResource
 Set-Alias -Name Create-MHProgressUpdateStream -Value New-MHProgressUpdateStream