+++ modules/AWSPowerShell/Cmdlets/AmazonMediaStoreData/Basic/Get-EMSDObjectMetadata-Cmdlet.cs
@@ -1,0 +1,155 @@
+/*******************************************************************************
+ *  Copyright 2012-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.MediaStoreData;
+using Amazon.MediaStoreData.Model;
+
+namespace Amazon.PowerShell.Cmdlets.EMSD
+{
+    /// <summary>
+    /// Gets the header for an object at the specified path.
+    /// </summary>
+    [Cmdlet("Get", "EMSDObjectMetadata")]
+    [OutputType("Amazon.MediaStoreData.Model.DescribeObjectResponse")]
+    [AWSCmdlet("Calls the AWS Elemental MediaStore Data Plane DescribeObject API operation.", Operation = new[] {"DescribeObject"})]
+    [AWSCmdletOutput("Amazon.MediaStoreData.Model.DescribeObjectResponse",
+        "This cmdlet returns a Amazon.MediaStoreData.Model.DescribeObjectResponse object containing multiple properties. The object can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class GetEMSDObjectMetadataCmdlet : AmazonMediaStoreDataClientCmdlet, IExecutor
+    {
+        
+        #region Parameter Path
+        /// <summary>
+        /// <para>
+        /// <para>The path (including the file name) where the object is stored in the container. Format:
+        /// &lt;folder name&gt;/&lt;folder name&gt;/&lt;file name&gt;</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
+        public System.String Path { get; set; }
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext
+            {
+                Region = this.Region,
+                Credentials = this.CurrentCredentials
+            };
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            context.Path = this.Path;
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new Amazon.MediaStoreData.Model.DescribeObjectRequest();
+            
+            if (cmdletContext.Path != null)
+            {
+                request.Path = cmdletContext.Path;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(context.Credentials, context.Region);
+            try
+            {
+                var response = CallAWSServiceOperation(client, request);
+                Dictionary<string, object> notes = null;
+                object pipelineOutput = response;
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response,
+                    Notes = notes
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.MediaStoreData.Model.DescribeObjectResponse CallAWSServiceOperation(IAmazonMediaStoreData client, Amazon.MediaStoreData.Model.DescribeObjectRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS Elemental MediaStore Data Plane", "DescribeObject");
+            try
+            {
+                #if DESKTOP
+                return client.DescribeObject(request);
+                #elif CORECLR
+                // todo: handle AggregateException and extract true service exception for rethrow
+                var task = client.DescribeObjectAsync(request);
+                return task.Result;
+                #else
+                        #error "Unknown build edition"
+                #endif
+            }
+            catch (AmazonServiceException exc)
+            {
+                var webException = exc.InnerException as System.Net.WebException;
+                if (webException != null)
+                {
+                    throw new Exception(Utils.Common.FormatNameResolutionFailureMessage(client.Config, webException.Message), webException);
+                }
+                throw;
+            }
+        }
+        
+        #endregion
+        
+        internal partial class CmdletContext : ExecutorContext
+        {
+            public System.String Path { get; set; }
+        }
+        
+    }
+}